chore: copy elision, moves innecesarios
