+++ client/renderables/hud/renderable_hud_health.cpp
@@ -2,7 +2,7 @@
 
 RenderableHUDHealth::RenderableHUDHealth(std::shared_ptr<AnimationProvider> animation_provider)
     : RenderableHUDBase(animation_provider) {
-    icon_animation = std::move(animation_provider->make_animation(get_icon_name()));
+    icon_animation = animation_provider->make_animation(get_icon_name());
 }
 
 std::string RenderableHUDHealth::get_icon_name() const {
