+++ client/renderables/hud/renderable_hud_money.cpp
@@ -6,7 +6,7 @@
 
 RenderableHUDMoney::RenderableHUDMoney(std::shared_ptr<AnimationProvider> animation_provider)
     : RenderableHUDBase(animation_provider) {
-    icon_animation = std::move(animation_provider->make_animation(get_icon_name()));
+    icon_animation = animation_provider->make_animation(get_icon_name());
 }
 
 std::string RenderableHUDMoney::get_icon_name() const {
